LayoutTests/media/modern-media-controls/buttons-container/buttons-container-buttons-property.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing <code>ButtonsContainer</code> buttons property.");

window.jsTestIsAsync = true;

const container = new ButtonsContainer({
    margin: 10,
    padding: 20
});

container.buttons = [new Button, new Button, new Button];

scheduler.frameDidFire = function()
{
    shouldBe("container.buttons.length", "3");
    shouldBe("container.children.length", "3");
    shouldBe("container.buttons", "container.children");
    finishMediaControlsTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
